pyproject.toml
[tool.poetry]
name = "multiply_backend_coding_task"
version = "0.1.0"
description = ""
authors = ["Michael Curtis <mike@multiply.ai>"]

[tool.poetry.dependencies]
python = "^3.9" # allows us to add flake8
Flask = "^2.1.2"
python-dotenv = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^24.4.1"
ruff = "^0.4.2"
pre-commit = "^3.7.0"
coverage = "^7.5.0"
pyhamcrest = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.isort]
profile = 'black'

[tool.black]
line-length = 88

[tool.coverage.run]
branch = true
command_line = '-m pytest'
source = ['multiply_backend_coding_task/']

[tool.coverage.report]
precision = 2
skip_covered = true
fail_under = 75
